replace existing modules
