+++ bundle/config.ts
@@ -1,10 +1,13 @@
-import * as Config from "#dist/effect/config/Config"
+import * as Config2 from "#dist/effect/config/Config2"
 import * as Effect from "#dist/effect/Effect"
+import * as Schema from "#dist/effect/schema/Schema"
 
-const config = Config.all({
-  API_KEY: Config.String("API_KEY"),
-  PORT: Config.Integer("PORT"),
-  LOCALHOST: Config.Url("LOCALHOST")
+const schema = Schema.Struct({
+  API_KEY: Schema.String,
+  PORT: Schema.Int,
+  LOCALHOST: Schema.URL
 })
+
+const config = Config2.schema(schema)
 
 Effect.runFork(config.asEffect())
