+++ packages/effect/src/config/index.ts
@@ -10,19 +10,4 @@ export * as Config from "./Config.ts"
 /**
  * @since 4.0.0
  */
-export * as Config2 from "./Config2.ts"
-
-/**
- * @since 4.0.0
- */
-export * as ConfigError from "./ConfigError.ts"
-
-/**
- * @since 4.0.0
- */
 export * as ConfigProvider from "./ConfigProvider.ts"
-
-/**
- * @since 4.0.0
- */
-export * as ConfigProvider2 from "./ConfigProvider2.ts"
